Add only-key option; Add @config key
For some use cases (bookmarking), the key does not require to have a
value which was considered as bad input until this PR. This PR adds the
ability to call a shortcut without a value. In this case the $TERM$ will
be replaced by an empty string.

Add @config key which will open the preferences page for the extension.

Update README and settings page to reflect on the new features and
examples.